fix mileage type

+++ src/main/java/com/faithjoyfundation/autopilotapi/v1/dto/car_managment/CarDTO.java
@@ -17,7 +17,7 @@ public class CarDTO {
 
     private String VIN;
 
-    private Double mileage;
+    private Integer mileage;
 
     private Integer year;
